Don't check dependencies when installing artifacts
This is currently causing issues randomly. We have no system in place which guarantees that all dependencies are available for all artifacts in the build_depends stage of the build process. With this change, this should no longer matter. This solution works, because the packages which have this issue are never needed to build the current package.