Update sqlite to 3.7.4 with different mirror
The tarball on the sqlite website was constantly changing, which is
totally unacceptable. The tag on the github should not change. I had to
adjust the patch a little bit to work with the next version, though. I
hope it works like expected, but I don't have an application I can test
this with.